pyproject.toml
[project]
name = "chainlit-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.1",
    "boto3>=1.40.2",
    "chainlit>=2.6.5",
    "contextvars>=2.4",
    "flask>=3.1.1",
    "flask-compress>=1.18",
    "playwright>=1.54.0",
    "pyjwt>=2.10.1",
    "python-engineio>=4.12.2",
    "python-socketio>=5.13.0",
    "redis>=6.3.0",
    "strands-agents-tools>=0.2.3",
    "uvicorn>=0.35.0",
]
